+++ cudax/test/stf/freeze/task_fence.cu
@@ -93,7 +93,7 @@ int main()
 
   ctx.finalize();
 
-  for (int i = 0; i < N; i++)
+  for (size_t i = 0; i < N; i++)
   {
     EXPECT(X[i] == 2 * X0(i) + 1);
   }
